refactor: split validation functionality from core services
