+++ services/offers/offers.service.js
@@ -2,7 +2,10 @@ const fs = require("fs/promises");
 
 class OfferService {
   static async getOffers() {
-    const file = await fs.readFile(__dirname + "/../data/offers.json", "utf8");
+    const file = await fs.readFile(
+      __dirname + "/../../data/offers.json",
+      "utf8"
+    );
     return JSON.parse(file);
   }
